convert money to big decimal

+++ app/src/main/java/ru/redbyte/exchangerate/data/balance/BalanceRepository.kt
@@ -11,21 +11,22 @@ import ru.redbyte.exchangerate.data.exchange.Currency
 import ru.redbyte.exchangerate.data.exchange.Currency.*
 import ru.redbyte.exchangerate.domain.balance.BalanceDataSource
 import java.lang.reflect.Type
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class BalanceRepository @Inject constructor(
         private val sharedPreferences: SharedPreferences,
         private val gson: Gson
 ) : BalanceDataSource {
-    private val balanceTypeToken: Type by lazy { typeTokenOf<HashMap<Currency, Double>>() }
+    private val balanceTypeToken: Type by lazy { typeTokenOf<HashMap<Currency, BigDecimal>>() }
 
-    override fun getBalance(): Single<Map<Currency, Double>> = Single.fromCallable {
+    override fun getBalance(): Single<Map<Currency, BigDecimal>> = Single.fromCallable {
         val balance = sharedPreferences.getString(BALANCE, "") ?: ""
         return@fromCallable gson.fromJson(balance, balanceTypeToken)
-                ?: mapOf(USD to 100.0, EUR to 100.0, GBP to 100.0, RUB to 100.0)
+                ?: mapOf(USD to BigDecimal(100.0), EUR to BigDecimal(100.0), GBP to BigDecimal(100.0), RUB to BigDecimal(100.0))
     }
 
-    override fun saveBalance(balance: Map<Currency, Double>): Completable = Completable.fromAction {
+    override fun saveBalance(balance: Map<Currency, BigDecimal>): Completable = Completable.fromAction {
         sharedPreferences.commit {
             putString(BALANCE, gson.toJson(balance))
         }
